add logo, move gCal to main list view, WIP: adding and styling footer
