+++ app/public/index.html
@@ -6,6 +6,7 @@
   <link rel="icon" href="%PUBLIC_URL%/got-it-logo-icon1.png" />
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css"
     integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
+  <link rel="stylesheet" type="text/css" href="../src/App.css" media="screen"/>
 
   <meta name="viewport" content="width=device-width, initial-scale=1" />
   <meta name="theme-color" content="#000000" />
@@ -15,7 +16,7 @@
 
   <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
 
-  <title>Got It!</title>
+  <title>Got It!></title>
 </head>
 
 <body>
@@ -32,4 +33,52 @@
     crossorigin="anonymous"></script>
 </body>
 
+<!-- Footer -->
+<footer class="footer text-center text-lg-start bg-light text-muted orange-background">
+
+  <!-- Section: Links  -->
+  <section class="orange-background">
+    <div class="orange-background text-center text-md-start mt-5">
+      <!-- Grid row -->
+      <div class="row mt-3 orange-background">
+        <!-- Grid column -->
+        <div class="col-md-3 col-lg-4 col-xl-3 mx-auto mb-4 pt-4 orange-background">
+          <!-- Content -->
+          <h6 class="text-uppercase fw-bold mb-6 " >
+            Got It!
+          </h6>
+          <p>
+            Developed as Meia Natividad's Final Project for Techtonica</p>
+          <p>May 2021 | Oakland, CA
+
+          </p>
+        </div>
+        <!-- Grid column -->
+
+        <!-- Grid column -->
+        <div class="col-md-4 col-lg-3 col-xl-3 mx-auto mb-md-0 mb-4 pt-4 orange-background" >
+          <!-- Links -->
+          <h6 class="text-uppercase fw-bold mb-6">
+            Contact
+          </h6>
+          <p class="mb-0">
+            <a href="mailto:meianatividad@gmail.com" target="_blank">Email Me</a>
+          </p>
+          <p class="mb-0">
+            <a href="https://github.com/m3ia/GotIt" target="_blank">Visit GitHub Repo</a>
+          </p>
+          <p class="mb-0">
+            <a href="https://linkedin.com/in/meiamay" target="_blank">Connect on LinkedIn</a>
+          </p>
+        </br>
+        </div>
+        <!-- Grid column -->
+      </div>
+      <!-- Grid row -->
+    </div>
+  </section>
+  <!-- Section: Links  -->
+
+</footer>
+<!-- Footer -->
 </html>
