+++ app/src/App.js
@@ -1,13 +1,12 @@
 import * as React from "react";
 import { Fragment, useState, useEffect } from "react";
 
-import gcal from "./ApiCalendar";
 import ListItems from "./ListItems";
 import ViewAllLists from "./ViewAllLists";
+import logo from "./got-it-logo1.png";
 // import * as apiClient from "./apiClient";
 
 const App = () => {
-  const [isAuthenticated, setIsAuthenticated] = useState(gcal.sign);
   const [page, setPage] = useState("home");
   const [selectedListId, setSelectedListId] = useState(null);
   const back = () => {
@@ -15,16 +14,20 @@ const App = () => {
     setSelectedListId(null);
   };
 
-  useEffect(() => {
-    gcal.onLoad(() => {
-      setIsAuthenticated(gcal.gapi.auth2.getAuthInstance().isSignedIn.get());
-      gcal.listenSign((sign) => setIsAuthenticated(sign));
-    });
-  }, []);
   return (
     <Fragment>
-      <Login {...{ isAuthenticated }} />
-      {isAuthenticated ? <Events /> : null}
+      <div className="jumbotron">
+        <div className="logo">
+          <img src={logo} className="logo" alt="Got It Logo" width="75%" />
+        </div>
+        <div className="description">
+          <p>
+            Got It! is the recurring to-do list for anyone who loves to prep and
+            plan on the daily, weekly, or monthly basis! With shareable lists,
+            Got It was made with families, friends, and collectives in mind.
+          </p>
+        </div>
+      </div>
       <div className="container">
         {page === "home" && (
           <ViewAllLists
@@ -39,31 +42,6 @@ const App = () => {
         )}
       </div>
     </Fragment>
-  );
-};
-
-const Login = ({ isAuthenticated }) =>
-  isAuthenticated ? (
-    <button onClick={gcal.handleSignoutClick}>Log out</button>
-  ) : (
-    <button onClick={gcal.handleAuthClick}>Log in</button>
-  );
-
-const Events = () => {
-  const [events, setEvents] = React.useState([]);
-
-  React.useEffect(() => {
-    gcal
-      .listUpcomingEvents(10)
-      .then(({ result: { items } }) => setEvents(items));
-  }, []);
-
-  return events.length === 0 ? null : (
-    <ul>
-      {events.map((event) => (
-        <li key={event.id}>{event.summary}</li>
-      ))}
-    </ul>
   );
 };
